LayoutTests/imported/w3c/web-platform-tests/html/rendering/replaced-elements/attributes-for-embedded-content-and-images/video-intrinsic-width-height-expected.txt


PASS no width/height attributes
FAIL only width attribute assert_equals: height expected 150 but got 50
FAIL only height attribute assert_equals: width expected 300 but got 200
PASS both width/height attributes
PASS both width/height attributes and style
FAIL both width/height attributes and style keeping aspect-ratio assert_equals: width expected 500 but got 300
